feat(current-offers): rename `current-promotions` to `current-offers`
